separated types declaration in new module
